+++ protocols/AXHARemoteUpdateProtocol.h
@@ -5,7 +5,7 @@
 
 @required
 
-- (void)updateProperty:(int)arg1 forDeviceID:(NSString *)arg2;
-- (void)writeValue:(id)arg1 forProperty:(int)arg2 forDeviceID:(NSString *)arg3;
+- (void)updateProperty:(unsigned long long)arg1 forDeviceID:(NSString *)arg2;
+- (void)writeValue:(id)arg1 forProperty:(unsigned long long)arg2 forDeviceID:(NSString *)arg3;
 
 @end
